refactor: decompose TUI input.rs into focused interaction modules [OPE-304] (#174)
* refactor: decompose TUI input.rs into focused interaction modules [OPE-304]

Split the 844-line monolithic input.rs into a module directory with clear
ownership boundaries:

- mod.rs (77 lines): main handle_key dispatcher/facade
- modal.rs (88 lines): secret input, provider select, model select handlers
- navigation.rs (137 lines): scroll helpers, messages panel, vim-style nav
- command_palette.rs (41 lines): command palette interaction
- setup.rs (17 lines): setup mode handler
- tests.rs (500 lines): complete test suite carried forward

Zero behavior change - public API (handle_key) remains unchanged.


* style: fix formatting in TUI input tests


---------
